cleaned up ui
