+++ react-client/src/Components/HeaderComponent.css
@@ -1,0 +1,11 @@
+.navbar-links li button {
+    font-family: "Poppins";
+    color: white;
+    text-decoration: none;
+    margin: 0 10px;
+    background: none;
+    border: none;
+    padding: 0;
+    cursor: pointer;
+    font-size: 16px;
+}
