--- react-client/src/Components/pages/TempWorkoutPage.css
@@ -1,6 +1,0 @@
-body {
-    background-color: white;
-    color: black;
-}
-
-/* Add any additional CSS rules here to override the current styles */
