+++ react-client/src/Components/pages/WorkoutPage.js
@@ -1,9 +1,8 @@
 import React from 'react';
 import WorkoutComponent from '../WorkoutComponent';
-//import './TempWorkoutPage.css'; //UNCOMMENT TO TEST THIS PAGE
+import './WorkoutPage.css';
 
 const WorkoutPage = () => {
-
     return (
         <div>
             <h1>Workout Tracking Page</h1>
